requirements.txt
openai
faiss-cpu
annoy
sentence_transformers
numpy
tiktoken